+++ post-trade/src/main/java/com/multiassetoms/posttrade/model/UnrealizedPnlException.java
@@ -1,0 +1,8 @@
+package com.multiassetoms.posttrade.model;
+
+public class UnrealizedPnlException extends RuntimeException {
+
+    public UnrealizedPnlException(String message) {
+        super(message);
+    }
+}